+++ CHANGELOG.md
@@ -7,13 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- NA
+- Fixed [Calc-ChemMassPercent.py] invalid variable name `Solid_@`
 
 ## [1.4.1] - 2024-01-04
 
 - Updated links
 - Added [Calc-ChemMassPercent.py] `--molar` option
-- Updated [Calc-ChemMassPercent.py] synopsis
 
 ## [1.4.0] - 2024-01-04